Sets the user password after user creation
The user is created without password and the password is set after that.
These two steps are needed when adding users to the samba4 sam database
directly without using command line tools. Also windows does this way
in Active Directory.

A new method 'modifications' allow to retrieve the modifications
made to the LDAP object in the last call to 'save'.